+++ server/src/bruFiles/completions/dictionaryBlocks/specificBlocks/getMethodBlockContentCompletions.ts
@@ -1,18 +1,10 @@
 import {
     MethodBlockKey,
     MethodBlockBodies,
-    MethodBlockAuthValues,
     Block,
     getMandatoryKeysForMethodBlock,
-    isAuthBlock,
-    AuthBlockName,
-    OAuth2AuthBlocksCommonKeys,
-    getMandatoryKeysForNonOAuth2Block,
-    AuthBlockNamesExcludingOAuth2,
     LineBreakType,
-    getExpectedAuthBlockForType,
     isBodyBlock,
-    getAuthTypesForNoDefinedAuthBlock,
     getBodyBlockTypeForNoDefinedBodyBlock,
 } from "@global_shared";
 import { LanguageFeatureBaseRequest } from "../../../../shared";
@@ -23,6 +15,7 @@ import { getTextEditForDictionaryBlockSimpleValue } from "../generic/getTextEdit
 import { CompletionItem } from "vscode-languageserver";
 import { getTextEditForInsertingBlock } from "../../../shared/getTextEditForInsertingBlock";
 import { getTextEditForRemovingBlock } from "../../../shared/getTextEditForRemovingBlock";
+import { getCompletionsForAuthTypeFieldValue } from "../generic/getCompletionsForAuthTypeFieldValue";
 
 export function getMethodBlockContentCompletions(
     request: LanguageFeatureBaseRequest,
@@ -69,7 +62,7 @@ function getValueCompletions(
     }
 
     if (isInAuthField) {
-        return getCompletionsForAuthFieldValue(
+        return getCompletionsForAuthTypeFieldValue(
             baseRequest,
             allBlocks,
             methodBlock,
@@ -83,76 +76,6 @@ function getValueCompletions(
         methodBlock,
         { currentLineContent: currentText, lineBreak },
     );
-}
-
-function getCompletionsForAuthFieldValue(
-    { documentHelper, position: { line } }: LanguageFeatureBaseRequest,
-    allBlocks: Block[],
-    methodBlock: Block,
-    content: { currentLineContent: string; lineBreak: LineBreakType },
-) {
-    const { currentLineContent, lineBreak } = content;
-    const existingAuthBlocks = allBlocks.filter(({ name }) =>
-        isAuthBlock(name),
-    );
-    const authTypesWithoutAnAuthBlock = getAuthTypesForNoDefinedAuthBlock();
-
-    if (existingAuthBlocks.length <= 1) {
-        return Object.values(MethodBlockAuthValues)
-            .map((authType) => {
-                const label = authType;
-                const textEdit = getTextEditForDictionaryBlockSimpleValue(
-                    line,
-                    currentLineContent,
-                    authType,
-                );
-
-                if (
-                    authTypesWithoutAnAuthBlock.includes(authType) &&
-                    existingAuthBlocks.length > 0
-                ) {
-                    return {
-                        label,
-                        textEdit,
-                        additionalTextEdits: [
-                            getTextEditForRemovingBlock(
-                                documentHelper,
-                                existingAuthBlocks[0],
-                            ),
-                        ].filter((val) => val != undefined),
-                    };
-                }
-
-                if (
-                    !authTypesWithoutAnAuthBlock.includes(authType) &&
-                    existingAuthBlocks.length == 0
-                ) {
-                    const newAuthBlockName =
-                        getExpectedAuthBlockForType(authType);
-                    const authBlockKeys =
-                        newAuthBlockName == AuthBlockName.OAuth2Auth
-                            ? Object.values(OAuth2AuthBlocksCommonKeys)
-                            : getMandatoryKeysForNonOAuth2Block(
-                                  newAuthBlockName as AuthBlockNamesExcludingOAuth2,
-                              );
-
-                    return {
-                        label,
-                        textEdit,
-                        additionalTextEdits: [
-                            getTextEditForInsertingBlock(methodBlock, {
-                                blockName: newAuthBlockName,
-                                content: { keys: authBlockKeys },
-                                lineBreak,
-                            }),
-                        ].filter((val) => val != undefined),
-                    };
-                }
-
-                return { label, textEdit };
-            })
-            .filter((val) => val != undefined);
-    }
 }
 
 function getCompletionsForBodyFieldValue(