+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Improve grouping and sorting of context menu items in collection explorer.
 - Provide completions for keys in all dictionary blocks.
 - Replace existing Snippets with context-aware completions for full blocks. Therefore, only blocks that would be valid are shown.
-- Add automatic updates of respective blocks for certain situations when changing value for auth or body field in a method block.
+    - The completions also provide additional text edits in certain cases where conditionally mandatory blocks either exist and would become invalid or don't exist and would become mandatory.
 
 ### Fixes / Minor Improvements